more improvements around video loading.
