+++ lib/src/features/video_details/presentation/widgets/video_player.dart
@@ -1,10 +1,10 @@
 import 'dart:math';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:peertuber/src/features/common/domain/entities/video.dart';
 import 'package:media_kit_video/media_kit_video.dart';
 import 'package:peertuber/src/features/common/presentation/bloc/media_player/media_player_bloc.dart';
+import 'package:peertuber/src/features/video_details/presentation/widgets/widgets.dart';
 
 class VideoPlayer extends StatelessWidget {
   final VideoEntity video;
@@ -111,94 +111,6 @@ class VideoPlayer extends StatelessWidget {
   }
 }
 
-class VideoSlider extends HookWidget {
-  const VideoSlider({
-    super.key,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    final position = useState<double>(0.0);
-    // Builder
-    return BlocBuilder<MediaPlayerBloc, MediaPlayerState>(
-      buildWhen: (previous, current) =>
-          [MediaPlayerPlaying, MediaPlayerPaused].contains(current.runtimeType),
-      builder: (context, state) {
-        if (state is MediaPlayerPlaying) {
-          final seekPosition = state.seekPosition;
-          var progress = ((seekPosition.inSeconds /
-                      context
-                          .read<MediaPlayerBloc>()
-                          .controller
-                          .player
-                          .state
-                          .duration
-                          .inSeconds) *
-                  100) *
-              0.01;
-
-          if (progress.isNaN || progress.isInfinite) {
-            progress = 0;
-          }
-
-          position.value = progress;
-        }
-
-        return Stack(
-          children: [
-            /*Slider(
-              value: progress,
-              max: 1,
-              min: 0,
-              onChanged: (value) => {},
-              onChangeEnd: (value) => player.seek(Duration(
-                  seconds: (value * player.state.duration.inSeconds).toInt())),
-            ),*/
-            LinearProgressIndicator(
-              value: position.value,
-              valueColor: AlwaysStoppedAnimation<Color>(
-                  Theme.of(context).colorScheme.primary),
-            ),
-          ],
-        );
-      },
-    );
-  }
-}
-
-class MiniPlayerControls extends StatelessWidget {
-  const MiniPlayerControls({
-    super.key,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return BlocBuilder<MediaPlayerBloc, MediaPlayerState>(
-      buildWhen: (previous, current) => [
-        MediaPlayerPlaying,
-        MediaPlayerPaused,
-      ].contains(current.runtimeType),
-      builder: (context, state) {
-        return IconButton(
-            onPressed: () {
-              state is MediaPlayerPlaying
-                  ? context
-                      .read<MediaPlayerBloc>()
-                      .add(PauseMedia(video: state.video))
-                  : (state is MediaPlayerPaused)
-                      ? context
-                          .read<MediaPlayerBloc>()
-                          .add(PlayMedia(video: state.video))
-                      : null;
-            },
-            icon: state is MediaPlayerPlaying
-                ? const Icon(Icons.pause)
-                : const Icon(Icons.play_arrow));
-      },
-    );
-  }
-}
-
 class _MiniPlayerVideoInfo extends StatelessWidget {
   final VideoEntity video;
   const _MiniPlayerVideoInfo({
@@ -267,67 +179,5 @@ class _VideoWidget extends StatelessWidget {
             fit: BoxFit.fitHeight);
       },
     );
-  }
-}
-
-class VideoPlayerControls extends HookWidget {
-  const VideoPlayerControls({
-    required this.state,
-    super.key,
-  });
-
-  final VideoState state;
-
-  @override
-  Widget build(BuildContext context) {
-    final showControls = useState<bool>(false);
-
-    return showControls.value
-        ? SizedBox(
-            width: double.infinity,
-            height: double.infinity,
-            child: Stack(
-              children: [
-                GestureDetector(
-                  onTap: () => showControls.value = !showControls.value,
-                  child: const AbsorbPointer(
-                    child: SizedBox(
-                      width: double.infinity,
-                      height: double.infinity,
-                    ),
-                  ),
-                ),
-                IconButton(
-                  onPressed: () =>
-                      context.read<MediaPlayerBloc>().add(MinimizePlayer()),
-                  iconSize: 30.0,
-                  icon: const Icon(Icons.keyboard_arrow_down),
-                ),
-                Center(
-                  child: IconButton(
-                    onPressed: () {
-                      state.widget.controller.player.playOrPause();
-                    },
-                    icon: StreamBuilder(
-                      stream: state.widget.controller.player.stream.playing,
-                      builder: (context, playing) => Icon(
-                        playing.data == true ? Icons.pause : Icons.play_arrow,
-                        size: 60,
-                      ),
-                    ),
-                  ),
-                ),
-              ],
-            ),
-          )
-        : GestureDetector(
-            onTap: () => showControls.value = !showControls.value,
-            child: const AbsorbPointer(
-              child: SizedBox(
-                width: double.infinity,
-                height: double.infinity,
-              ),
-            ),
-          );
   }
 }
