+++ lib/src/features/video_details/presentation/widgets/widgets.dart
@@ -1,2 +1,6 @@
 export 'video_info.dart';
 export 'video_player.dart';
+export 'video_slider.dart';
+export 'video_player_overlay.dart';
+export 'mini_player_controls.dart';
+export 'video_search_results.dart';
